Updates
Okay, so hear me out if we put the functions and data in a class, we might run into the problem of having multiple classes for the user.  By putting everything in a namespace it is not only a perfect Singleton (we cannot run into that problem), but we can move the functions inside to whatever files we want (although I suppose we could also use partial classes for that).